+++ sim/src/make/spawner.rs
@@ -294,120 +294,3 @@ impl TripSpec {
         })
     }
 }
-
-/// Specifies where a trip begins or ends.
-#[derive(Serialize, Deserialize, PartialEq, Eq, PartialOrd, Ord, Debug, Clone, Copy)]
-pub enum TripEndpoint {
-    Bldg(BuildingID),
-    Border(IntersectionID),
-    /// Used for interactive spawning, tests, etc. For now, only valid as a trip's start.
-    SuddenlyAppear(Position),
-}
-
-impl TripEndpoint {
-    /// Figure out a single PathRequest that goes between two TripEndpoints. Assume a single mode
-    /// the entire time -- no walking to a car before driving, for instance. The result probably
-    /// won't be exactly what would happen on a real trip between the endpoints because of this
-    /// assumption.
-    pub fn path_req(
-        from: TripEndpoint,
-        to: TripEndpoint,
-        mode: TripMode,
-        map: &Map,
-    ) -> Option<PathRequest> {
-        let start = from.pos(mode, true, map)?;
-        let end = to.pos(mode, false, map)?;
-        Some(match mode {
-            TripMode::Walk | TripMode::Transit => PathRequest::walking(start, end),
-            TripMode::Bike => PathRequest::vehicle(start, end, PathConstraints::Bike),
-            // Only cars leaving from a building might turn out from the driveway in a special way
-            TripMode::Drive => {
-                if matches!(from, TripEndpoint::Bldg(_)) {
-                    PathRequest::leave_from_driveway(start, end, PathConstraints::Car, map)
-                } else {
-                    PathRequest::vehicle(start, end, PathConstraints::Car)
-                }
-            }
-        })
-    }
-
-    fn start_sidewalk_spot(&self, map: &Map) -> Result<SidewalkSpot> {
-        match self {
-            TripEndpoint::Bldg(b) => Ok(SidewalkSpot::building(*b, map)),
-            TripEndpoint::Border(i) => SidewalkSpot::start_at_border(*i, map)
-                .ok_or_else(|| anyhow!("can't start walking from {}", i)),
-            TripEndpoint::SuddenlyAppear(pos) => Ok(SidewalkSpot::suddenly_appear(*pos, map)),
-        }
-    }
-
-    fn end_sidewalk_spot(&self, map: &Map) -> Result<SidewalkSpot> {
-        match self {
-            TripEndpoint::Bldg(b) => Ok(SidewalkSpot::building(*b, map)),
-            TripEndpoint::Border(i) => SidewalkSpot::end_at_border(*i, map)
-                .ok_or_else(|| anyhow!("can't end walking at {}", i)),
-            TripEndpoint::SuddenlyAppear(_) => unreachable!(),
-        }
-    }
-
-    fn driving_goal(&self, constraints: PathConstraints, map: &Map) -> Result<DrivingGoal> {
-        match self {
-            TripEndpoint::Bldg(b) => Ok(DrivingGoal::ParkNear(*b)),
-            TripEndpoint::Border(i) => map
-                .get_i(*i)
-                .some_incoming_road(map)
-                .and_then(|dr| {
-                    let lanes = dr.lanes(constraints, map);
-                    if lanes.is_empty() {
-                        None
-                    } else {
-                        // TODO ideally could use any
-                        Some(DrivingGoal::Border(dr.dst_i(map), lanes[0]))
-                    }
-                })
-                .ok_or_else(|| anyhow!("can't end at {} for {:?}", i, constraints)),
-            TripEndpoint::SuddenlyAppear(_) => unreachable!(),
-        }
-    }
-
-    fn pos(self, mode: TripMode, from: bool, map: &Map) -> Option<Position> {
-        match mode {
-            TripMode::Walk | TripMode::Transit => (if from {
-                self.start_sidewalk_spot(map)
-            } else {
-                self.end_sidewalk_spot(map)
-            })
-            .ok()
-            .map(|spot| spot.sidewalk_pos),
-            TripMode::Drive | TripMode::Bike => {
-                if from {
-                    match self {
-                        // Fall through and use DrivingGoal also to start.
-                        TripEndpoint::Bldg(_) => {}
-                        TripEndpoint::Border(i) => {
-                            return map.get_i(i).some_outgoing_road(map).and_then(|dr| {
-                                dr.lanes(mode.to_constraints(), map)
-                                    .get(0)
-                                    .map(|l| Position::start(*l))
-                            });
-                        }
-                        TripEndpoint::SuddenlyAppear(pos) => {
-                            return Some(pos);
-                        }
-                    }
-                }
-                self.driving_goal(mode.to_constraints(), map)
-                    .ok()
-                    .and_then(|goal| goal.goal_pos(mode.to_constraints(), map))
-            }
-        }
-    }
-
-    /// Returns a point representing where this endpoint is.
-    pub fn pt(&self, map: &Map) -> Pt2D {
-        match self {
-            TripEndpoint::Bldg(b) => map.get_b(*b).polygon.center(),
-            TripEndpoint::Border(i) => map.get_i(*i).polygon.center(),
-            TripEndpoint::SuddenlyAppear(pos) => pos.pt(map),
-        }
-    }
-}
